classes/recipes: Switch virtual/XXX-gcc to virtualk/XXX-cc
The idea of the base class dependency is to say "yes, I need a C cross compiler"
and this was never meant to be gcc specific. The clang layer understadable tries
to change this dependency as it would be confusing otherwise.

Whilst this will cause some pain, change the default to "cc" from "gcc" to make
it clear what is really wanted here.

At the same time, change the much less widely used "g++" to "c++" for similar
reasons.

Backwards compatibility mappings could be left but are just going to confuse
things in future so we'll just require users to update. This will make multiple
toolchain support much cleaner in the future.

(From OE-Core rev: 5c31ab9b016d2d97b39c5d15eccae6981d57d905)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>