bitbake: fetch2: gitsm: remove useless need_update check
The need_update check is useless to decide if the clone directory or
temporary directory should be used to parse the submodules. If the git
repository needs an update, it does not help to unpack the git archive
and use it instead of the clone directory. Furthermore, it is useless
to call the function of the class itself because it bypasses the check
itself.

(Bitbake rev: 37a9194a3076bce44c6b148143ef4d25c4695bd1)

Signed-off-by: Stefan Herbrechtsmeier <stefan.herbrechtsmeier@weidmueller.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>