devtool: ide-sdk deploy-target without bitbake
The main goal of devtool ide-sdk is to use bitbake for generating
SDKs and consistent IDE configurations for working on the source code of
one or more recipes in workspaces created by devtool modify.
When the IDE is configured for a build system such as CMake or Meson,
the IDE must not invoke bitbake. But bitbake can be called explicitly
whenever something on the recipes changes and the image, the sysroots
or the IDE configuration needs to be updated.

A cross development workflow usually requires the following steps with
clear separation of responsibilities between the IDE and bitbake:
- Setup the complete build environment                 (bitbake-setup)
  - Get the layers:
  - Initialize the build configuration
- Choose some recipes to work on                      (devtool modify)
- Setup the image and the SDK   (devtool ide-sdk, shared sstate-cache)
- Work on the recipe in the workspace                            (IDE)
  - (re-)configure
  - compile
  - run unit tests (on the host with Qemu)
  - remote debuggging on the target device which includes
    - do_install                                          (IDE,pseudo)
    - deploying the artifacts to the target device        (IDE,pseudo)
    - start gdbserver on the target device
    - start GDB on the host device
- Update the SDK and image                          (bitbake, devtool)

For the configure, compile and unit test tasks there is already full
IDE integration, without any call to bitbake. But for the remote
debugging or more precisely for the deployment to the target device,
bitbake gets called which has several disadvantages:

- Bitbake runs a work queue for all tasks of the recipe. There is no
  way to run only the do_install task. Bitbake always tries to find
  out if a previous task needs to be executed and automatically does
  so. Since bitbake has no information about which steps have already
  been executed by the IDE, there is a great chance that bitbake does
  much more than necessary. It could also happen that bitbake ignores
  changes done by the developer in the IDE and overrules some manual
  steps. This leads to a very bad user experience. Therefore it is
  really important to delegate all steps for the application
  development workflow to the IDE and use bitbake only for providing
  and updating the SDK and the image.
  An example of this is that bitbake recompiles everything when a file
  listed in the CONFIGURE_FILES variable is changed. This makes no
  sense in the IDE context and can be extremely tedious.
  The complexity of resolving many dependencies is simply not part of
  an application development workflow. So there is no argument for
  invoking a powerful but heavy tool. Bitbake adds complexity that
  application developers usually don't want to understand or debug.
  As soon as bitbake is needed for daily work, most application
  developers have a defensive attitude towards the overall solution.
  Parsing all recipes, creating task queues and running tasks which
  do not have to be executed makes it very slow, which is another
  major disadvantage.
- bitbake -T (mem-res server mode) could avoid re-parsing. But when
  bitbake is needed to update the SDK or build an image, reparsing
  all recipes is essential. Since inotify has been removed from
  bitbake, the -T option is by design no longer usable to run a
  bitbake server as an IDE backend.
  Starting bitbake with -T requires manually stopping bitbake
  whenever reparsing is expected. That's not intuitive at all.
- Bitbake -b is very quick but ignores bbappends. This is a bug which
  probably could be fixed. But it does not look like the solution for
  the overall problem here anyway.

Therefore a much better user experience can be achieved if all tasks
listed above are calls to a build tool such as cmake or self-contained
scripts which are under control of the IDE. This commit refactors the
code to no longer invoke bitbake for the target-deploy step. The script
which is used by the IDE for the deployment of the artifacts gets
improved to become a self-contained script which uses pseudo but does
not call bitbake anymore.

The script re-implements some bitbake internals and needs to be kept in
sync with bitbake which is not perfect. But refactoring bitbake's code
to become reusable for this script as well looks not feasible at all.

(From OE-Core rev: ae21c40e4392d3f25f8c69d1b92f7d0e159f737c)

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>